# ui/desktop/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "daisy-server",
    "description": "An AI agent",
    "contact": {
      "name": "Block",
      "email": "ai-oss-tools@block.xyz"
    },
    "license": {
      "name": "Apache-2.0"
    },
    "version": "1.8.0"
  },
  "paths": {
...
    "/confirm": {
      "post": {
        "tags": [
          "super::routes::reply"
        ],
        "operationId": "confirm_permission",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PermissionConfirmationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Permission action is confirmed",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid secret key"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PermissionConfirmationRequest": {
        "type": "object",
        "required": [
          "id",
          "action",
          "session_id"
        ],
        "properties": {
          "action": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "principal_type": {
            "$ref": "#/components/schemas/PrincipalType"
          },
          "session_id": {
            "type": "string"
          }
        }
      },
...
      "PrincipalType": {
        "type": "string",
        "enum": [
          "Extension",
          "Tool"
        ]
      },
...
    }
  }
}